+++ CddWep/views/main/sopClause/sopClause.js
@@ -24,35 +24,55 @@ define(function (require) {
         //初始化
         $scope.searchData = {};
         //获取分页数据
-        var currentCheck = function (page, callback) {
-            console.log($scope.searchData);
-            $http.post(url + '/sop/showPageList', $.extend({}, page, $scope.searchData)).success(callback);
-        };
-        $scope.projectItem = app.get('Paginator').list(currentCheck, 6);
-        console.log($scope.projectItem);
+        if (role == 1) {
+            var currentCheck = function (page, callback) {
+                console.log($scope.searchData);
+                $http.post(url + '/sop/showPageList', $.extend({}, page, $scope.searchData)).success(callback);
+            };
+            $scope.projectItem = app.get('Paginator').list(currentCheck, 6);
+            console.log($scope.projectItem);
 
-        //编辑事件
-        $scope.edit = function (item) {
-            item = JSON.stringify(item);
-            $location.path('/main/sopClause/newSopClause/' + item);
-        };
+            //下载
+            $scope.download = function (fileName) {
+                window.location.href = url + '/file/download?path=' + fileName;
+            };
+        }else if(role == 2){
+            var currentCheck = function (page, callback) {
+                console.log($scope.searchData);
+                $http.post(url + '/sop/showPageList', $.extend({}, page, $scope.searchData)).success(callback);
+            };
+            $scope.projectItem = app.get('Paginator').list(currentCheck, 6);
+            console.log($scope.projectItem);
+
+            //下载
+            $scope.download = function (fileName) {
+                window.location.href = url + '/file/download?path=' + fileName;
+            };
+        }else if(role == 3){
+            var currentCheck = function (page, callback) {
+                console.log($scope.searchData);
+                $http.post(url + '/sop/showPageList', $.extend({}, page, $scope.searchData)).success(callback);
+            };
+            $scope.projectItem = app.get('Paginator').list(currentCheck, 6);
+            console.log($scope.projectItem);
+
+            //编辑事件
+            $scope.edit = function (item) {
+                item = JSON.stringify(item);
+                $location.path('/main/sopClause/newSopClause/' + item);
+            };
+            //删除
+            $scope.deleteById = function (id) {
+                $http.get(url + '/sop/delete?id=' + id).success(function () {
+                    $scope.projectItem = app.get('Paginator').list(currentCheck, 6);
+                    yMake.layer.msg('删除成功！', {icon: '1', time: 2000});
+                }).error(function(){
+                    yMake.layer.msg('删除失败！', {icon: '2', time: 2000});
+                })
+            };
+        }
 
 
-
-        //删除
-        $scope.deleteById = function (id) {
-            $http.get(url + '/sop/delete?id=' + id).success(function () {
-                $scope.projectItem = app.get('Paginator').list(currentCheck, 6);
-                yMake.layer.msg('删除成功！', {icon: '1', time: 2000});
-            }).error(function(){
-                yMake.layer.msg('删除失败！', {icon: '2', time: 2000});
-            })
-        };
-
-        //下载
-        $scope.download = function (fileName) {
-            window.location.href = url + '/file/download?path=' + fileName;
-        };
 
         /*$scope.items = [{company:'652856',types:'入库操作SOP',time:'2017-10-22',shim:'关于XXXX的通报'},
          {company:'6528561',types:'类型2',time:'2017-10-22',shim:'关于XXXX的通报'},
